Juego (casi) listo para publicar
A falta de reviciones menores el juego ya estaría listo para pasarselo a
un grupo de personas y empezar a tomar datos.

+++ Juego/Assets/Scripts/SocketController.cs
@@ -163,7 +163,19 @@ public class SocketController : MonoBehaviour
     //Mato el proceso de pythom
     void OnApplicationQuit()
     {
-        p.Kill();
+        UnityEngine.Debug.Log("process-Kill on Quit");
+        Process[] pname = Process.GetProcessesByName("python");
+        if (pname.Length != 0)
+            p.Kill();
+    }
+
+    //
+    void OnDestroy()
+    {
+        UnityEngine.Debug.Log("process-Kill on Destroy");
+        Process[] pname = Process.GetProcessesByName("python");
+        if (pname.Length != 0)
+            p.Kill();
     }
